Issue #43: Save the last download result in a preference.
In the preference activity for the speed test, we'll display
the size of the last download, instead of downloading the
file each time.

+++ src/ca/rmen/android/networkmonitor/app/speedtest/SpeedTestUpload.java
@@ -69,7 +69,7 @@ public class SpeedTestUpload {
             long after = System.currentTimeMillis();
             ftp.logout();
             ftp.disconnect();
-            return new SpeedTestResult((int) uploadConfig.file.length(), after - before, SpeedTestStatus.SUCCESS);
+            return new SpeedTestResult(uploadConfig.file.length(), after - before, SpeedTestStatus.SUCCESS);
         } catch (SocketException e) {
             Log.e(TAG, "upload " + e.getMessage(), e);
             return new SpeedTestResult(0, 0, SpeedTestStatus.FAILURE);
